backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.1
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-dotenv==1.0.1
pydantic==2.9.2
gunicorn==21.2.0

